Small Changes
Reworked the folders to be more organized. Added inventory (TODO: fix
the text). Added a stub for Relics. Must implement them somehow. I gotta
think more about that.

Next on the agenda is making buildings keep track of what enemies are in
it and only having the enemies attack you when you are in the building.

I got engr to study for though, so I don't know when I can get to this.
Maybe tomorrow.

MAIN CLASSES GOTTA BE FINISHED BY NEXT WEEK! THAT INCLUDES QUEST CRAP!