Prayer Implementation
Finished up prayers. The mechanics for relics also changed. If a relic
does something passive like increase attack, you should add that in the
equip function to modify the player class. You should undo the change in
dequip. The Defend method is still up for if you need a relic to do
retribution damage or something.